Update Optimizer sample programs
Change sample names as needed, correct code and comments

Remove WebOptimizedPDF sample, update sample copyright dates and
description.
