+++ DotNET/Sample_Source/DocumentOptimization/PDFOptimize/Properties/AssemblyInfo.cs
@@ -5,12 +5,12 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("PDFOptimizer")]
+[assembly: AssemblyTitle("PDFOptimize")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("PDFOptimizer")]
-[assembly: AssemblyCopyright("Copyright ©  2008-2016")]
+[assembly: AssemblyProduct("PDFOptimize")]
+[assembly: AssemblyCopyright("Copyright ©  2008-2022")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
